Jam.py:
Bug with open method, when order_by parameter is an empty list fixed
Bug with sql generated when default order is set and fields parameter do not contain any of field of default order in the open method fixed
The exception handling of errors occuring in the code, when inplace editor is used reworked so developer can find the reson of an error

Administrator:

Tabs are now created for opened modules.

Demo application:

Search for catalogs reworked.