Jam.py:
Import functionality for SQLITE databases is available now
Creation of foreign field indexes for SQLITE databases are removed
Popovers for fields with help attribute reworked

Admin:

Bugs related to tabs are fixed

Demo:

For Customers item email sending functionlally is added. It demonstrates the use of server method to execute script defined in the server module from client module, how to use fields of item with virtual_table to create a form for input of data.