Library:
History of changes made by users can now be stored. See Saving the history of changes made by users
Local filtering of dataset records is reworked and published. See Filtered, on_filter_record
clone method is published
Application is now throws an exception when an attempt is made to get or set a value to a field when the dataset is empty.

Application Builder:

Delete reports after attribute is added to Project parameters
Some changes to interface are made.