Synchronization of parameters and reloading of the task tree when metadata changes for web applications running on parallel processes reworked.
Import of metadata reworked.

Created the ability to import metadata from the migration folder when the server is restarted.

Migration to another database is available now. See How to migrate to another database

virtual_table is now a read-only property on the client and server. For an item which virtual_table property is true, calling the open method creates an empty data set, and calling the apply method does nothing.

When importing a table the virtual_table attribute id read only now.

title_line_count option is added to the table_options specifies the number of lines of text displayed in a title row, if it is 0, the height of the row is determined by the contents of the title cells It can be set in Application Builder.